+++ packages/article-in-depth/src/article-in-depth.js
@@ -31,6 +31,7 @@ class ArticleInDepth extends Component {
       backgroundColour,
       byline,
       flags,
+      hasVideo,
       headline,
       label,
       publicationName,
@@ -44,6 +45,7 @@ class ArticleInDepth extends Component {
         <ArticleHeader
           backgroundColour={backgroundColour}
           flags={flags}
+          hasVideo={hasVideo}
           headline={getHeadline(headline, shortHeadline)}
           label={label}
           standfirst={standfirst}