+++ packages/article-magazine-comment/src/article-magazine-comment.web.js
@@ -28,6 +28,7 @@ class ArticlePage extends Component {
       },
       byline,
       flags,
+      hasVideo,
       headline,
       label,
       publicationName,
@@ -42,6 +43,7 @@ class ArticlePage extends Component {
           authorImage={author.image}
           byline={byline}
           flags={flags}
+          hasVideo={hasVideo}
           headline={getHeadline(headline, shortHeadline)}
           label={label}
           publicationName={publicationName}