+++ packages/article-magazine-standard/src/article-header/article-header-prop-types.js
@@ -3,6 +3,7 @@ import PropTypes from "prop-types";
 const articleHeaderPropTypes = {
   byline: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
   flags: PropTypes.arrayOf(PropTypes.string),
+  hasVideo: PropTypes.bool,
   headline: PropTypes.string.isRequired,
   label: PropTypes.string,
   publicationName: PropTypes.string.isRequired,
@@ -12,6 +13,7 @@ const articleHeaderPropTypes = {
 
 const articleHeaderDefaultProps = {
   flags: null,
+  hasVideo: false,
   label: null,
   standfirst: null
 };