+++ packages/provider-queries/src/article.js
@@ -4,7 +4,16 @@ import gql from "graphql-tag";
 export default addTypenameToDocument(gql`
   query ArticleQuery($id: ID!) {
     article(id: $id) {
+      backgroundColour {
+        rgba {
+          red
+          green
+          blue
+          alpha
+        }
+      }
       content
+      dropcapsDisabled
       flags
       keywords
       leadAsset {
@@ -52,6 +61,14 @@ export default addTypenameToDocument(gql`
       }
       standfirst
       template
+      textColour {
+        rgba {
+          red
+          green
+          blue
+          alpha
+        }
+      }
       topics(maxCount: 5) {
         name
         slug