+++ packages/provider-queries/src/article.web.js
@@ -4,8 +4,17 @@ import gql from "graphql-tag";
 export default addTypenameToDocument(gql`
   query ArticleQuery($id: ID!) {
     article(id: $id) {
+      backgroundColour {
+        rgba {
+          red
+          green
+          blue
+          alpha
+        }
+      }
       commentsEnabled
       content: paywalledContent
+      dropcapsDisabled
       flags
       keywords
       leadAsset {
@@ -53,6 +62,14 @@ export default addTypenameToDocument(gql`
       }
       standfirst
       template
+      textColour {
+        rgba {
+          red
+          green
+          blue
+          alpha
+        }
+      }
       topics(maxCount: 5) {
         name
         slug