+++ packages/utils/src/gql-rgba-to-style.js
@@ -1,0 +1,6 @@
+export default colour =>
+  colour && colour.rgba
+    ? `rgba(${colour.rgba.red}, ${colour.rgba.green}, ${colour.rgba.blue}, ${
+        colour.rgba.alpha
+      })`
+    : null;