+++ packages/article-flag/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.17](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.2.16...@times-components/article-flag@2.2.17) (2019-01-14)
+
+**Note:** Version bump only for package @times-components/article-flag
+
+
+
+
+
 ## [2.2.16](https://github.com/newsuk/times-components/compare/@times-components/article-flag@2.2.15...@times-components/article-flag@2.2.16) (2019-01-08)
 
 **Note:** Version bump only for package @times-components/article-flag