+++ packages/article/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.5.36](https://github.com/newsuk/times-components/compare/@times-components/article@6.5.35...@times-components/article@6.5.36) (2019-01-14)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [6.5.35](https://github.com/newsuk/times-components/compare/@times-components/article@6.5.34...@times-components/article@6.5.35) (2019-01-11)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [6.5.34](https://github.com/newsuk/times-components/compare/@times-components/article@6.5.33...@times-components/article@6.5.34) (2019-01-09)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [6.5.33](https://github.com/newsuk/times-components/compare/@times-components/article@6.5.32...@times-components/article@6.5.33) (2019-01-09)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [6.5.32](https://github.com/newsuk/times-components/compare/@times-components/article@6.5.31...@times-components/article@6.5.32) (2019-01-08)
 
 **Note:** Version bump only for package @times-components/article